Added hardware_counter feature flag and removed old metric structs from the metric crate
