+++ Cargo.toml
@@ -26,12 +26,14 @@ exclude = [
 ]
 
 [features]
-default = ["serde"]
+default = ["serde", "metrics", "hardware_counters"]
 arcon_serde = ["serde_json", "bincode", "serde", "uuid/serde"]
 unsafe_flight = ["abomonation", "abomonation_derive", "arcon_macros/unsafe_flight"]
 kafka = ["rdkafka", "futures", "serde_json", "serde"]
 thread_pinning = ["kompact/thread_pinning"]
 socket = ["tokio-util", "futures", "serde_json", "serde"]
+arcon_metrics=["metrics", "metrics-exporter-prometheus", "metrics-util"]
+hardware_counters=["perf-event", "arcon_metrics"]
 #metrics = []
 
 # state features
@@ -64,13 +66,10 @@ twox-hash = "1.5.0"
 prost = "0.7"
 prost-types = "0.7"
 bytes = "1.0"
-metrics = "0.16.0"
-metrics-exporter-prometheus = "0.5.0"
-metrics-util = "0.9.1"
+
 
 arrow = "4.3"
 datafusion = "4.0"
-perf-event = "0.4.7"
 
 # Optional
 rdkafka = { version = "0.26", optional = true }
@@ -83,6 +82,15 @@ bincode = { version = "1.2.1", optional = true }
 rayon = { version = "1.3.0", optional = true }
 abomonation = { version = "0.7.3", optional = true }
 abomonation_derive = { version = "0.5.0", optional = true }
+
+
+metrics = {version="0.16.0", optional=true}
+metrics-exporter-prometheus = {version="0.5.0", optional=true}
+metrics-util = {version="0.9.1", optional=true}
+
+#[target.'cfg(target_os = "linux")'.dependencies]
+perf-event = { version = "0.4.7", optional = true }
+
 
 [dev-dependencies]
 tempfile = "3"
