+++ guide/examples/src/bin/kafka_source.rs
@@ -1,19 +1,15 @@
 use arcon::prelude::*;
 
-
 fn main() {
-
     let mut perf_events = PerfEvents::new();
     perf_events.add(HardwareCounter::CPU_CYCLES);
+    perf_events.add(HardwareCounter::CACHE_MISSES);
 
     let consumer_conf = KafkaConsumerConf::default()
         .with_topic("test")
         .set("group.id", "test")
         .set("bootstrap.servers", "localhost:9092")
         .set("enable.auto.commit", "false");
-
-
-
 
     let mut pipeline = Pipeline::default()
         .kafka(consumer_conf, JsonSchema::new(), 1, |conf| {
@@ -23,7 +19,6 @@ fn main() {
         .operator(OperatorBuilder {
             constructor: Arc::new(|_| Map::new(|x| x + 10)),
             conf: OperatorConf {
-                // parallelism_strategy: ParallelismStrategy::Managed,
                 perf_events,
                 ..Default::default()
             },
